darko_marketing/requirements.txt
# LLM providers
anthropic>=0.25.0
openai>=1.30.0

# Core utilities
python-dotenv>=1.0.0
rich>=13.7.0
pydantic>=2.7.0
httpx>=0.27.0

# Live intel scraping
praw>=7.7.0

# Distribution — X (Twitter)
tweepy>=4.14.0

# Distribution — Ghost CMS
PyJWT>=2.8.0
